fix: update error handling for loading wdio config file (#58)
## Proposed changes

Fixed error handling when a configuration file is incomplete (e.g., has
syntax errors).
Fixed to delete the hierarchy under the subordinate SPEC file when the
configuration file is changed to be incomplete, since it is not deleted.


Added Smoke test for this case as well.

## Types of changes

[//]: # 'What types of changes does your code introduce to WebdriverIO?'
[//]: # '_Put an `x` in the boxes that apply_'

- [ ] Polish (an improvement for an existing feature)
- [X] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improvements to the project's docs)
- [ ] Internal updates (everything related to internal scripts,
governance documentation and CI files)

## Checklist

[//]: # "_Put an `x` in the boxes that apply. You can also fill these
out after creating the PR. If you're unsure about any of them, don't
hesitate to ask. We're here to help! This is simply a reminder of what
we are going to look for before merging your code._"

- [X] I have read the
[CONTRIBUTING](https://github.com/webdriverio/vscode-webdriverio/blob/main/CONTRIBUTION.md)
doc
- [X] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added the necessary documentation (if appropriate)
- [ ] I have added proper type definitions for new commands (if
appropriate)

## Further comments

[//]: # 'If this is a relatively large or complex change, kick off the
discussion by explaining why you chose the solution you did and what
alternatives you considered, etc...'

### Reviewers: @webdriverio/project-committers