feat: simplification of the startup process (#70)
## Proposed changes

Change the startup process for feature extensions.

With the initial implementation for worker process startup (controlling
the startup process with a semaphore) and the fix for #62 (starting the
worker if it was stopped when accessed), there is no longer a need to
pre-start all workers.

Simplify the logic for the extension startup process and remove
unnecessary source code.

Additionally, add E2E tests related to changes in the ConfigFilePattern
in Settings.

## Types of changes

[//]: # 'What types of changes does your code introduce to WebdriverIO?'
[//]: # '_Put an `x` in the boxes that apply_'

- [X] Polish (an improvement for an existing feature)
- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improvements to the project's docs)
- [ ] Internal updates (everything related to internal scripts,
governance documentation and CI files)

## Checklist

[//]: # "_Put an `x` in the boxes that apply. You can also fill these
out after creating the PR. If you're unsure about any of them, don't
hesitate to ask. We're here to help! This is simply a reminder of what
we are going to look for before merging your code._"

- [X] I have read the
[CONTRIBUTING](https://github.com/webdriverio/vscode-webdriverio/blob/main/CONTRIBUTING.md)
doc
- [X] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added the necessary documentation (if appropriate)
- [ ] I have added proper type definitions for new commands (if
appropriate)

## Further comments

[//]: # 'If this is a relatively large or complex change, kick off the
discussion by explaining why you chose the solution you did and what
alternatives you considered, etc...'

### Reviewers: @webdriverio/project-committers